+++ Src/@Types/index.ts
@@ -1,4 +1,5 @@
 export { KiraErrorCode } from "./common";
+
 export type {
   PresenceStatus,
   BorderAlign,
@@ -18,6 +19,7 @@ export type {
   MemberEventDrawOverrides,
   AttachmentOptions,
 } from "./common";
+
 export type {
   KiraThemeName,
   KiraThemePalette,
@@ -25,6 +27,7 @@ export type {
   ColorResolutionSources,
   CanvasBadge,
 } from "./theme";
+
 export type {
   KiraBadge,
   KiraNameplate,
@@ -46,6 +49,7 @@ export type {
   CatalogEntry,
   RawProfileBadge,
 } from "./user";
+
 export type {
   RankOptions,
   MemberEventOptions,
@@ -70,6 +74,7 @@ export type {
   GiveawayOptions,
   GiveawayLayout,
 } from "./options";
+
 export type {
   MoonlinkTrackLike,
   LavalinkTrackLike,